model_hub/requirements.txt
Cerberus==1.3.5
coloredlogs==15.0.1
filelock==3.9.0
flatbuffers==23.5.26
fsspec==2023.4.0
h5py==3.10.0
humanfriendly==10.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.26.1
onnx==1.14.1
onnxruntime-training==1.16.0
onnxruntime-training-cpu==0.1
packaging==23.2
protobuf==3.20.3
sympy==1.12
torch==2.1.0+cpu
typing_extensions==4.4.0
